Add dpl_mail
 For handling mail related features and configuration.
 Initially the module handles the azure environment variable and
 transforms it into a configuration array that the azure mailer module
can consume.
