+++ web/modules/custom/dpl_mail/dpl_mail.info.yml
@@ -1,0 +1,9 @@
+name: "DPL Mail"
+description: "Handles mail related features and configuration"
+package: DPL
+
+type: module
+core_version_requirement: ^10
+
+dependencies:
+  - azure_mailer:azure_mailer
